fix comment and move slash handler's tests to slash_test.go
